Change semantics of positional shapes depending on pjit nesting
One interesting angle of pjit is that it is a boundary between a multi-controller
world in which `.shape` attributes of all arrays (and avals) correspond to
slices of data that are internall to a given process, and a single-controller
world where `.shape` refers to the global array constructed by concatenating
per-device chunks. I haven't fully appreciated this previously which made
pjit nests (and xmaps in pjits) to incorrectly increase shapes with every
level of nesting, when only the outermost call that should make the change.

We now keep track of a flag that determines whether the positional shape of
avals we see is global or local in any given context. Note that sizes of named
axes have been and still are global only.

PiperOrigin-RevId: 397165574